requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-allauth==0.44.0
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-modeltranslation==0.17.1
django-recaptcha3==0.4.0
filelock==3.0.12
h11==0.12.0
httptools==0.2.0
idna==2.10
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.5
virtualenv==20.4.2
